+++ dialog/cocoa/dlg.go
