+++ dialog/cocoa/dlg.h
