+++ dialog/cocoa/dlg.m
