+++ dialog/dlgs.go
