+++ dialog/util.go
